// src/categories/types.ts
export type ResponseCategory = {
    category_id: string;
    name: string;
    nick_name: string;
    product: boolean;
    created_at: string;
    updated_at: string;
};

export type ResultCategory = {
    categoryId: string;
    name: string;
    nickName: string;
    product: boolean;
    createdAt: string;
    updatedAt: string;
};